+++ internal/xsql/bad_conn.go
@@ -1,0 +1,12 @@
+//go:build !go1.18
+// +build !go1.18
+
+package xsql
+
+import (
+	"database/sql/driver"
+)
+
+func badConn(err error) error {
+	return driver.ErrBadConn
+}