+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Implemented `database/sql` driver over `ydb-go-sdk`
 * Refactored session lifecycle in session pool for fix flaked `TestTable`
 * Fixed deadlock in topicreader batcher, while add and read raw server messages
 * Fixed bug in `db.Topic()` with send response to stop partition message